pyproject.toml
[tool.poetry]
name = "beatmebackend"
version = "0.1.0"
description = "let's fight"
authors = ["Ilia Snezhkov <iasnezhkov@gmail.com>"]

readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
aioredis = "^1.3.1"
tortoise-orm = {extras = ["accel"], version = "^0.16.14"}
asyncpg = "^0.20.1"
uvicorn = "^0.11.5"
pyJWT = "^1.7.1"
orjson = "^3.2.1"
httpx = "^0.13.3"
python-decouple = "^3.3"
typer = "^0.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.0.0"
pytest-asyncio = "^0.14.0"
mypy = "^0.782"
pylint = "^2.5.3"
pytest-cov = "^2.10.0"
requests = "^2.24.0"
# wemake-python-styleguide = "^0.14.0"

[tool.isort]
line_length = 88
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
multi_line_output = 3
use_parentheses = true
not_skip = "__init__.py"
skip_glob = ["*/setup.py"]
filter_files = true

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

